Revert "fix: rename v2 to fault_proving_extra_storage"
This reverts commit 4560027d740a5e815192f7df66a658228ca05cb1.
