+++ crates/fuel-core/src/graphql_api/storage/da_compression/v2/address.rs
